MAE-174 - Rename the RichTextWidgetOptions to RichTextWidgetConfiguration - move the public JS (inline editor) API
