+++ Kentico.Widget.RichText/Kentico.Widget.RichText.csproj
@@ -530,7 +530,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Controllers\KenticoRichTextWidgetController.cs" />
-    <Compile Include="Models\Widgets\RichTextWidgetOptions.cs" />
+    <Compile Include="Models\Widgets\RichTextWidgetConfiguration.cs" />
     <Compile Include="PageLink\JsonCamelCaseResult.cs" />
     <Compile Include="Models\Widgets\RichTextWidgetProperties.cs" />
     <Compile Include="Models\Widgets\RichTextWidgetViewModel.cs" />
