--- Kentico.Widget.RichText/Models/Widgets/RichTextWidgetOptions.cs
@@ -1,15 +1,0 @@
-﻿using Kentico.Components.Web.Mvc.Widgets.Controllers;
-
-namespace Kentico.Components.Web.Mvc.Widgets.Models
-{
-    /// <summary>
-    /// Represents configurable options of the rich text widget.
-    /// </summary>
-    public class RichTextWidgetOptions
-    {
-        /// <summary>
-        /// Configuration name for the underlying inline editor.
-        /// </summary>
-        public string ConfigurationName { get; set; } = KenticoRichTextWidgetController.IDENTIFIER;
-    }
-}
